fix complexity
