+++ src/backend/libs/efiling-cso-starter/src/main/java/ca/bc/gov/open/jag/efilingcsostarter/config/AutoConfiguration.java
@@ -8,7 +8,7 @@ import ca.bc.gov.open.jag.efilingcommons.model.EfilingSoapClientProperties;
 import ca.bc.gov.open.jag.efilingcommons.model.SoapProperties;
 import ca.bc.gov.open.jag.efilingcommons.service.EfilingAccountService;
 import ca.bc.gov.open.jag.efilingcommons.service.EfilingDocumentService;
-import ca.bc.gov.open.jag.efilingcsostarter.CSODocumentServiceImpl;
+import ca.bc.gov.open.jag.efilingcsostarter.CsoDocumentServiceImpl;
 import ca.bc.gov.open.jag.efilingcsostarter.CsoAccountServiceImpl;
 import ca.bc.gov.open.jag.efilingcsostarter.mappers.AccountDetailsMapper;
 import ca.bc.gov.open.jag.efilingcsostarter.mappers.AccountDetailsMapperImpl;
@@ -67,7 +67,7 @@ public class AutoConfiguration {
     @Bean
     @ConditionalOnMissingBean({EfilingDocumentService.class})
     public EfilingDocumentService efilingDocumentService(FilingStatusFacadeBean filingStatusFacadeBean) {
-        return new CSODocumentServiceImpl(filingStatusFacadeBean);
+        return new CsoDocumentServiceImpl(filingStatusFacadeBean);
     }
 
     public <T> T getPort(Clients clients, Class<T> type) {
